Revert "split user state from auth state, and reduce overhead on auth state"
This reverts commit 4df93abef09832c04502e334a0ff0958328b097b.
