+++ lib/redux/authentication/actions.dart
@@ -1,11 +1,13 @@
 import 'dart:convert';
 
+import 'package:flutter_redux_navigation/flutter_redux_navigation.dart';
 import 'package:http/http.dart' as http;
 import 'package:redux/redux.dart';
 import 'package:redux_thunk/redux_thunk.dart';
 
 import 'package:scibble/models/auth_pkce.dart';
 import 'package:scibble/models/token.dart';
+import 'package:scibble/models/user.dart';
 import 'package:scibble/redux/store.dart';
 
 class SetAuthPKCE {
@@ -16,6 +18,11 @@ class SetAuthPKCE {
 class SetToken {
   final Token payload;
   SetToken(this.payload);
+}
+
+class SetUser {
+  final User payload;
+  SetUser(this.payload);
 }
 
 ThunkAction<AppState> tradeCodeForToken = (Store<AppState> store) async {
@@ -37,4 +44,15 @@ ThunkAction<AppState> tradeCodeForToken = (Store<AppState> store) async {
     Token token = Token.fromJson(json.decode(response.body));
     store.dispatch(SetToken(token));
   }
+};
+
+ThunkAction<AppState> getUserProfile = (Store<AppState> store) async {
+  final token = store.state.auth.tokenState.token;
+  var response = await http.get(
+    'https://online.ntnu.no/api/v1/profile/',
+    headers: {'Authorization': 'Bearer ${token.accessToken}'},
+  );
+  var user = User.fromJson(json.decode(response.body));
+  await store.dispatch(SetUser(user));
+  await store.dispatch(NavigateToAction.replace('/home'));
 };
