+++ lib/redux/authentication/reducer.dart
@@ -1,18 +1,31 @@
-import 'package:scibble/models/token.dart';
 import 'package:scibble/redux/authentication/actions.dart';
 import 'package:scibble/redux/authentication/state.dart';
+import 'package:scibble/redux/store.dart';
 
 AuthenticationState authenticationReducer(
     AuthenticationState state, dynamic action) {
-  return new AuthenticationState(
-    token: tokenReducer(state.token, action),
-    authPKCEState: authPKCEReducer(state.authPKCEState, action),
-  );
+  if (action is Logout) {
+    return new AuthenticationState.initialState(
+        state.authPKCEState.pkce.clientId);
+  } else {
+    return new AuthenticationState(
+      userState: userReducer(state.userState, action),
+      tokenState: tokenReducer(state.tokenState, action),
+      authPKCEState: authPKCEReducer(state.authPKCEState, action),
+    );
+  }
 }
 
-Token tokenReducer(Token prevState, dynamic action) {
+UserState userReducer(UserState prevState, dynamic action) {
+  if (action is SetUser) {
+    return new UserState(user: action.payload);
+  }
+  return prevState;
+}
+
+TokenState tokenReducer(TokenState prevState, dynamic action) {
   if (action is SetToken) {
-    return action.payload;
+    return new TokenState(token: action.payload);
   }
   return prevState;
 }
