+++ lib/redux/authentication/state.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:scibble/models/auth_pkce.dart';
 import 'package:scibble/models/token.dart';
+import 'package:scibble/models/user.dart';
 
 @immutable
 class AuthPKCEState {
@@ -8,18 +9,33 @@ class AuthPKCEState {
   const AuthPKCEState({@required this.pkce});
 }
 
+@immutable
+class TokenState {
+  final Token token;
+  const TokenState({@required this.token});
+}
+
+@immutable
+class UserState {
+  final User user;
+  const UserState({@required this.user});
+}
+
 class AuthenticationState {
   final AuthPKCEState authPKCEState;
-  final Token token;
+  final TokenState tokenState;
+  final UserState userState;
 
   AuthenticationState({
     this.authPKCEState,
-    this.token,
+    this.tokenState,
+    this.userState,
   });
 
   factory AuthenticationState.initialState(String clientId) =>
       AuthenticationState(
         authPKCEState: new AuthPKCEState(pkce: new AuthPKCE(clientId)),
-        token: null,
+        tokenState: new TokenState(token: null),
+        userState: new UserState(user: null),
       );
 }
